# src/config/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/usuarios": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "example": "any"
                },
                "apellido": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/usuarios-texto": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/usuarios-html": {
      "get": {
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/usuarios/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "example": "any"
                },
                "apellido": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "openapi": {
      "type": "string",
      "example": "3.0.0"
    },
    "info": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "API de Usuarios"
        },
        "version": {
          "type": "string",
          "example": "1.0.0"
        },
        "description": {
          "type": "string",
          "example": "Documentación de la API de Usuarios"
        }
      }
    },
    "servers": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "http://localhost:3000"
          },
          "description": {
            "type": "string",
            "example": "Servidor de desarrollo"
          }
        }
      }
    },
    "components": {
      "type": "object",
      "properties": {
        "securitySchemes": {
          "type": "object",
          "properties": {
            "bearerAuth": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "http"
                },
                "scheme": {
                  "type": "string",
                  "example": "bearer"
                }
              }
            }
          }
        }
      }
    }
  },
  "apis": [
    "./src/controllers/usuarios/usuarios.ts",
    "./src/routes/usuarios/usuarios.ts",
    "./src/models/usuarios/usuarios.ts"
  ]
}